Restructure models into separate submodules
