+++ modnet/matbench/benchmark.py
@@ -56,9 +56,9 @@ def matbench_benchmark(
         fit_settings: Any settings to pass to `model.fit(...)` directly
             (typically when not performing hyperparameter optimisation).
         classification: Whether all tasks are classification rather than regression.
-        model_type: whether to use "MODNetModel" or "Ensemble_MODNetModel" for benchmarking.
-            Ensemble_MODNetModel will additionally provide the "stds" inside the result dict.
-        n_models: number of inner models for "Ensemble_MODNetModel" (if used). Note  that
+        model_type: whether to use "MODNetModel" or "EnsembleMODNetModel" for benchmarking.
+            EnsembleMODNetModel will additionally provide the "stds" inside the result dict.
+        n_models: number of inner models for "EnsembleMODNetModel" (if used). Note  that
             if hp_optimization is set to True this value will be overwritten by fit_preset (125).
         save_folds: Whether to save dataframes with pre-processed fold
             data (e.g. feature selection).
@@ -167,7 +167,7 @@ def train_fold(
     fold_ind, (train_data, test_data) = fold
 
     results = {}
-    from modnet.models import MODNetModel, Ensemble_MODNetModel
+    from modnet.models import MODNetModel, EnsembleMODNetModel
     if classification:
         fit_settings["num_classes"] = {t: 2 for t in target_weights}
 
@@ -190,15 +190,15 @@ def train_fold(
             target_weights,
             **model_settings
         )
-    elif model_type == "Ensemble_MODNetModel":
-        model = Ensemble_MODNetModel(
+    elif model_type == "EnsembleMODNetModel":
+        model = EnsembleMODNetModel(
             target,
             target_weights,
             n_models=n_models,
             **model_settings
         )
     else:
-        raise RuntimeError(f"{model_type} not supported. Please choose from \"MODNetModel\" or \"MODNetModel\".")
+        raise RuntimeError(f"{model_type} not supported. Please choose from \"MODNetModel\" or \"EnsembleMODNetModel\".")
 
     if hp_optimization:
         models, val_losses, best_learning_curve, learning_curves, best_presets = model.fit_preset(
@@ -235,12 +235,12 @@ def train_fold(
 
     try:
         if classification:
-            if model_type == "Ensemble_MODNetModel":
+            if model_type == "EnsembleMODNetModel":
                 predictions, stds = model.predict(test_data, return_prob=True, return_unc=True)
             else:
                 predictions = model.predict(test_data, return_prob=True)
         else:
-            if model_type == "Ensemble_MODNetModel":
+            if model_type == "EnsembleMODNetModel":
                 predictions, stds = model.predict(test_data, return_unc=True)
             else:
                 predictions = model.predict(test_data)
@@ -280,7 +280,7 @@ def train_fold(
         df_test.to_csv("folds/test_f{}.csv".format(ind + 1))
 
     results["predictions"] = predictions
-    if model_type == "Ensemble_MODNetModel":
+    if model_type == "EnsembleMODNetModel":
         results["stds"] = stds
     results["targets"] = targets
     results["errors"] = errors
