+++ modnet/models/__init__.py
@@ -1,0 +1,5 @@
+from .vanilla import MODNetModel
+from .bayesian import BayesianMODNetModel
+from .ensemble import EnsembleMODNetModel
+
+__all__ = ("MODNetModel", "BayesianMODNetModel", "EnsembleMODNetModel")
